+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "pkg/ui/docs/themes/LoveIt"]
 	path = pkg/ui/docs/themes/LoveIt
 	url = https://github.com/dillonzq/LoveIt.git
+[submodule "third_party/googleapis"]
+	path = third_party/googleapis
+	url = https://github.com/googleapis/googleapis.git
