+++ docs/env.md
@@ -32,28 +32,18 @@ File generated by cli. DO NOT EDIT.
 | "STORE_RETHINKDB_URI" | localhost:28015 | RethinkDB URI |
 | "STORE_SCYLLA_URI" | localhost:9042 | Scylla URI |
 | "STORE_SQLITE_PATH" | /tmp/links.sqlite | SQLite URI |
-| "LOG_LEVEL" | logger.INFO_LEVEL |  |
-| "LOG_TIME_FORMAT" | time.RFC3339Nano |  |
 | "TRACER_SERVICE_NAME" | ShortLink | Service Name |
 | "TRACER_URI" | localhost:6831 | Tracing addr:host |
-| "MQ_ENABLED" | false | Enabled MQ-service |
-| "SENTRY_DSN" |  | key for sentry |
-| "GRPC_SERVER_PORT" | 50051 | gRPC port |
-| "GRPC_SERVER_CERT_PATH" | ops/cert/shortlink-server.pem | gRPC server cert |
-| "GRPC_SERVER_KEY_PATH" | ops/cert/shortlink-server-key.pem | gRPC server key |
-| "GRPC_CLIENT_PORT" | 50051 | gRPC port |
-| "GRPC_CLIENT_CERT_PATH" | ops/cert/intermediate_ca.pem | gRPC client cert |
 | "LOG_LEVEL" | logger.INFO_LEVEL |  |
 | "LOG_TIME_FORMAT" | time.RFC3339Nano |  |
+| "SENTRY_DSN" |  | key for sentry |
+| "MQ_ENABLED" | false | Enabled MQ-service |
 | "TRACER_SERVICE_NAME" | ShortLink |  |
 | "TRACER_URI" | localhost:6831 |  |
-| "MQ_ENABLED" | false |  |
+| "LOG_LEVEL" | logger.INFO_LEVEL |  |
+| "LOG_TIME_FORMAT" | time.RFC3339Nano |  |
 | "SENTRY_DSN" |  |  |
-| "GRPC_SERVER_PORT" | 50051 |  |
-| "GRPC_SERVER_CERT_PATH" | ops/cert/shortlink-server.pem |  |
-| "GRPC_SERVER_KEY_PATH" | ops/cert/shortlink-server-key.pem |  |
-| "GRPC_CLIENT_PORT" | 50051 |  |
-| "GRPC_CLIENT_CERT_PATH" | ops/cert/intermediate_ca.pem |  |
+| "MQ_ENABLED" | false |  |
 | "STORE_TYPE" | ram | Select: postgres, mongo, mysql, redis, dgraph, sqlite, leveldb, badger, ram, scylla, cassandra |
 | "MQ_TYPE" | rabbitmq | Select: kafka, rabbitmq, nats |
 | "MQ_KAFKA_URI" | localhost:9092 | Kafka URI |
@@ -62,3 +52,12 @@ File generated by cli. DO NOT EDIT.
 | "API_TYPE" | http-chi | Select: http-chi, gRPC-web, graphql, cloudevents, go-kit |
 | "API_PORT" | 7070 | API port |
 | "API_TIMEOUT" | 60 | Request Timeout |
+| "LOG_LEVEL" | logger.INFO_LEVEL |  |
+| "LOG_TIME_FORMAT" | time.RFC3339Nano |  |
+| "LOG_LEVEL" | logger.INFO_LEVEL |  |
+| "LOG_TIME_FORMAT" | time.RFC3339Nano |  |
+| "GRPC_SERVER_PORT" | 50051 | gRPC port |
+| "GRPC_SERVER_CERT_PATH" | ops/cert/shortlink-server.pem | gRPC server cert |
+| "GRPC_SERVER_KEY_PATH" | ops/cert/shortlink-server-key.pem | gRPC server key |
+| "GRPC_CLIENT_PORT" | 50051 | gRPC port |
+| "GRPC_CLIENT_CERT_PATH" | ops/cert/intermediate_ca.pem | gRPC client cert |
