+++ docs/shortctl.md
@@ -19,4 +19,4 @@ shortctl [flags]
       --o string             Output file path (default "./docs/env.md")
 ```
 
-###### Auto generated by spf13/cobra on 26-Sep-2020
+###### Auto generated by spf13/cobra on 14-Dec-2020
