+++ internal/metadata/infrastructure/rpc/rpc.go
@@ -1,3 +1,5 @@
+//go:generate protoc -I../../domain --go_out=Minternal/metadata/domain/rpc.proto=.:. --go-grpc_out=Minternal/metadata/domain/rpc.proto=.:. --go_opt=paths=source_relative --go-grpc_opt=paths=source_relative rpc.proto
+
 /*
 Metadata Service. Infrastructure layer
 */
@@ -6,19 +8,13 @@ package rpc
 import (
 	"context"
 
-	"github.com/batazor/shortlink/internal/di"
 	"github.com/batazor/shortlink/internal/logger"
 	"github.com/batazor/shortlink/internal/metadata/application"
-	rpc "github.com/batazor/shortlink/internal/metadata/domain"
 	meta_store "github.com/batazor/shortlink/internal/metadata/infrastructure/store"
+	"github.com/batazor/shortlink/pkg/rpc"
 )
 
-type MetadataServer struct {
-	service *application.Service
-	log     logger.Logger
-}
-
-func New(runRPCServer *di.RPCServer, st *meta_store.MetaStore, log logger.Logger) (*MetadataServer, error) {
+func New(runRPCServer *rpc.RPCServer, st *meta_store.MetaStore, log logger.Logger) (*MetadataServer, error) {
 	server := MetadataServer{
 		// Create Service Application
 		service: &application.Service{
@@ -27,36 +23,33 @@ func New(runRPCServer *di.RPCServer, st *meta_store.MetaStore, log logger.Logger
 		log: log,
 	}
 
-	service := &rpc.MetadataService{
-		Get: server.Get,
-		Set: server.Set,
-	}
-
 	// Register services
-	rpc.RegisterMetadataService(runRPCServer.Server, service)
+	RegisterMetadataServer(runRPCServer.Server, server)
 	runRPCServer.Run()
 
 	return &server, nil
 }
 
-func (m *MetadataServer) Get(ctx context.Context, req *rpc.GetMetaRequest) (*rpc.GetMetaResponse, error) {
+func (m *MetadataServer) Get(ctx context.Context, req *GetMetaRequest) (*GetMetaResponse, error) {
 	meta, err := m.service.Get(ctx, req.Id)
 	if err != nil {
 		return nil, err
 	}
 
-	return &rpc.GetMetaResponse{
+	return &GetMetaResponse{
 		Meta: meta,
 	}, nil
 }
 
-func (m *MetadataServer) Set(ctx context.Context, req *rpc.SetMetaRequest) (*rpc.SetMetaResponse, error) {
+func (m *MetadataServer) Set(ctx context.Context, req *SetMetaRequest) (*SetMetaResponse, error) {
 	meta, err := m.service.Set(ctx, req.Id)
 	if err != nil {
 		return nil, err
 	}
 
-	return &rpc.SetMetaResponse{
+	return &SetMetaResponse{
 		Meta: meta,
 	}, nil
 }
+
+func (m *MetadataServer) mustEmbedUnimplementedMetadataServer() {}
